feat: move layer handling to separate module
